en construction

+++ View/user/index.html.php
@@ -1,6 +1,7 @@
 <?php
     use creepy\Controller\AbstractController;
-    use creepy\Model\Entity\User;
+use creepy\Controller\UserController;
+use creepy\Model\Entity\User;
     use creepy\Model\Entity\Role;
     use creepy\Model\Manager\UserManager;
 
@@ -22,12 +23,47 @@
                <div id="divMail">
                    <p>Email : <span><?=UserManager::getUserById($_SESSION['user']->getId())->getEmail() ?></span></p>
                </div>
+
+            </div>
+            <div class="articleForm">
+                <?php
+                if (UserController::verifyRole()) {?>
+                    <h3 >Ajoutez un article</h3>
+                    <form action="/index.php?c=user" method="post" id="">
+                        <div>
+                            <input type="text" placeholder="Titre de l'article">
+                        </div>
+                        <div>
+                            <textarea name="" id="" cols="30" rows="10" placeholder="Texte de l'article"></textarea>
+                        </div>
+                        <div>
+                            <input type="submit" value="créer" name="save" class="save">
+                        </div>
+                    </form>
+                    <?php
+                } else { ?>
+                <?php }
+                ?>
+
+
             </div>
         </div>
         <div id="user-article">
             <div>
-                <div>MES ARTICLE</div>
-                <div></div>
+                <h2 >Règlement</h2>
+                <div id="regles">
+                    <p>
+                        Bonjour à tous et à toutes, par ce présent règlement nous vous informons que quelques règles sont
+                        en vigueur sur ce site.
+                    </p>
+                    <p>
+                        premierement assurez-vous de ne donnée a personnes vos information personnel tel que :
+                    </p>
+                    <p>
+                        mot de passe
+                    </p>
+                </div>
+
 
 
                 <br><br><br><br><br>
@@ -35,19 +71,26 @@
 
 
             </div>
-            <div>
-                <h3 style="text-align: center">User</h3>
+            <div id="userList">
                 <?php
-                foreach ($data['users_list'] as $user) {
-                    /* @var User $user */ ?>
+                if (UserController::verifyRole()) {?>
+                    <h3>User List</h3>
+                    <?php
+                    foreach ($data['users_list'] as $user) {
+                        /* @var User $user */ ?>
                         <div>
-                            <div style="display: flex; justify-content: space-between">
+                            <div>
                                 <a href="/index.php?c=user&a=show-user&id<?= $user->getId() ?>"><?= $user->getPseudo() ?></a>
                                 <a href="/index.php?c=user&a=delete-user&id<?= $user->getId() ?>">Supprimer</a>
                             </div>
                         </div>
-                     <?php
-                } ?>
+                        <?php
+                    } ?>
+                    <?php
+                } else { ?>
+                <?php }
+                ?>
+
             </div>
         </div>
     </div>
